# Conflicts:
#	public/_admin/css/app.css
#	public/css/app.css
#	public/mix-manifest.json
#	resources/views/admin/livewire/products-list/products-list.blade.php

+++ database/seeders/BrandsTableSeeder.php
@@ -3,23 +3,23 @@
 namespace Database\Seeders;
 
 use Domain\Products\Models\Brand;
-use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 
-class BrandsTableSeeder extends Seeder
+class BrandsTableSeeder extends BaseSeeder
 {
     /**
      * Run the database seeds.
      *
      * @return void
-     * @throws \Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\DiskDoesNotExist
-     * @throws \Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\FileDoesNotExist
-     * @throws \Spatie\MediaLibrary\Exceptions\FileCannotBeAdded\FileIsTooBig
      */
     public function run()
     {
+        if (!$this->shouldClearData()) {
+            return;
+        }
+
         DB::statement('SET FOREIGN_KEY_CHECKS=0;');
         Brand::query()->truncate();
         DB::statement('SET FOREIGN_KEY_CHECKS=1;');
